fs: increase zone extent size to 64 bits
Increase the zone extent size to be 64 bits in stead of 32, to
support very large file sizes in large zones.

This change is not backward-compatible with the current on-disk
format, so we need to step up the version.

+++ fs/io_zenfs.h
@@ -29,10 +29,10 @@ namespace ROCKSDB_NAMESPACE {
 class ZoneExtent {
  public:
   uint64_t start_;
-  uint32_t length_;
+  uint64_t length_;
   Zone* zone_;
 
-  explicit ZoneExtent(uint64_t start, uint32_t length, Zone* zone);
+  explicit ZoneExtent(uint64_t start, uint64_t length, Zone* zone);
   Status DecodeFrom(Slice* input);
   void EncodeTo(std::string* output);
   void EncodeJson(std::ostream& json_stream);
